Updated Lyron roles.

+++ index.js
@@ -71,6 +71,8 @@ bot.on("message", (message) => {
         message.reply("Roles not set. Contact Admin")
         return
       }
+      var botRole = server.roles.cache.find(r => r.name == "Lyron")
+      if (!botRole) return
 
       let oldChannel = server.channels.cache.find(c => c.name == name && c.type == "text")
       if (oldChannel) oldChannel.delete()
@@ -94,6 +96,10 @@ bot.on("message", (message) => {
               },
               {
                 id: role.id,
+                allow: ['VIEW_CHANNEL']
+              },
+              {
+                id: botRole.id,
                 allow: ['VIEW_CHANNEL']
               }
             ])
